feat: 회원 초대 수락 검증 추가
회원의 상태가 waiting인지 확인

+++ src/main/java/com/kanvan/team/service/TeamService.java
@@ -98,7 +98,7 @@ public class TeamService {
         Member waitingMember = memberRepository.findById(inviteId).orElseThrow(
                 () -> new CustomException(ErrorCode.MEMBER_NOT_FOUND));
 
-        if (waitingMember.getMember().equals(user)) {
+        if (waitingMember.getMember().equals(user) && waitingMember.getInviteStatus() == Invite.WAITING) {
             if (request.getInvite() == Invite.REFUSE) {
                 memberRepository.delete(waitingMember);
             } else {
